refactor: Add toDouble() method in sayur_extension.dart
This commit adds a new method, toDouble(), to the StringExtension class in sayur_extension.dart. The toDouble() method parses a string and converts it to a double value. This method is useful for converting string representations of numbers to their corresponding double values.

Refactor the code to include the new toDouble() method in the StringExtension class.